Refactor AddLocationMemoFragment, MapFragment
