+++ moti/app/src/main/res/values/strings.xml
@@ -6,4 +6,5 @@
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="query_hint">검색하기</string>
     <string name="recent_search">최근 검색</string>
+    <string name="delete_memo">삭제하기</string>
 </resources>
